requirements.txt
pdfplumber==0.11.4
Pillow>=10.0.0
google-generativeai==0.7.2
elevenlabs==1.10.0
requests>=2.31.0

google-cloud-texttospeech==2.20.0

python-dotenv==1.0.1
